Refactor class repository
- Add abstract class
- Add spec for abstract method
