+++ lib/active_record/sharding/shard_repository.rb
@@ -1,6 +1,6 @@
 module ActiveRecord
   module Sharding
-    class ShardRepository
+    class ShardRepository < AbstractRepository
       attr_reader :base_class
 
       def initialize(cluster_config, base_class)
@@ -22,24 +22,6 @@ module ActiveRecord
       end
 
       private
-
-      def generate_model_for_shard(connection_name)
-        base_class_name = @base_class.name
-        class_name = generate_class_name connection_name
-
-        model = Class.new(base_class) do
-          self.table_name = base_class.table_name
-
-          module_eval <<-RUBY, __FILE__, __LINE__ + 1
-                  def self.name
-                    "#{base_class_name}::#{class_name}"
-                  end
-                RUBY
-        end
-
-        model.class_eval { establish_connection(connection_name) }
-        model
-      end
 
       def generate_class_name(connection_name)
         "ShardFor#{connection_name.to_s.tr('-', '_').classify}"
