+++ lib/activerecord-sharding.rb
@@ -5,6 +5,7 @@ require "active_record/sharding/errors"
 require "active_record/sharding/config"
 require "active_record/sharding/cluster_config"
 require "active_record/sharding/modulo_router"
+require "active_record/sharding/abstract_repository"
 require "active_record/sharding/shard_repository"
 require "active_record/sharding/database_tasks"
 require "active_record/sharding/model"
